Extract project settings modal state
